[#5] Event refactoring to match GLFW API /

+++ waker-online/andromeda/source/core/system/events/mouse.hpp
@@ -7,13 +7,13 @@ namespace Andromeda {
     namespace Event {
         class Mouse_Move : public Event {
           public:
-            Mouse_Move(int32_t x, int32_t y) : m_Mouse_X(x), m_Mouse_Y(y) {}
+            Mouse_Move(int x, int y) : m_Mouse_X(x), m_Mouse_Y(y) {}
 
-            inline float get_x() const {
+            inline double get_x() const {
                 return m_Mouse_X;
             }
 
-            inline float get_y() const {
+            inline double get_y() const {
                 return m_Mouse_Y;
             }
 
@@ -27,18 +27,18 @@ namespace Andromeda {
             EVENT_CLASS_CATEGORY(Event_Category_Input | Event_Category_Mouse)
 
           private:
-            int32_t m_Mouse_X, m_Mouse_Y;
+            int m_Mouse_X, m_Mouse_Y;
         };
 
         class Mouse_Scroll : public Event {
           public:
-            Mouse_Scroll(int32_t x_Offset, int32_t y_Offset) : m_X_Offset(x_Offset), m_Y_Offset(y_Offset) {}
+            Mouse_Scroll(int x_Offset, int y_Offset) : m_X_Offset(x_Offset), m_Y_Offset(y_Offset) {}
 
-            inline int32_t get_x_offset() const {
+            inline int get_x_offset() const {
                 return m_X_Offset;
             }
 
-            inline int32_t get_y_offset() const {
+            inline int get_y_offset() const {
                 return m_Y_Offset;
             }
 
@@ -51,7 +51,7 @@ namespace Andromeda {
             EVENT_CLASS_TYPE(Mouse_Scroll)
             EVENT_CLASS_CATEGORY(Event_Category_Input | Event_Category_Mouse)
           private:
-            int32_t m_X_Offset, m_Y_Offset;
+            int m_X_Offset, m_Y_Offset;
         };
 
         class Mouse_Button : public Event {
