+++ waker-online/andromeda/source/core/system/input/input.hpp
@@ -8,9 +8,9 @@ namespace Andromeda {
       public:
         static bool is_key_pressed(Key_Code key);
         static bool is_mouse_button_pressed(Mouse_Code mouse_button);
-        static std::pair<float, float> get_mouse_position();
-        static float get_mouse_x();
-        static float get_mouse_y();
+        static std::pair<double, double> get_mouse_position();
+        static double get_mouse_x();
+        static double get_mouse_y();
     };
 
 } /* Andromeda */
