+++ waker-online/andromeda/source/core/system/platform/linux/input.cpp
@@ -6,29 +6,29 @@
 namespace Andromeda {
     bool Input::is_key_pressed(Key_Code key) {
         auto window = static_cast<GLFWwindow *>(Instance::get_instance().get_window().get_native_window());
-        auto state = glfwGetKey(window, static_cast<int32_t>(key));
+        auto state = glfwGetKey(window, static_cast<int>(key));
         return state == GLFW_PRESS || state == GLFW_REPEAT;
     }
 
     bool Input::is_mouse_button_pressed(Mouse_Code button) {
         auto window = static_cast<GLFWwindow *>(Instance::get_instance().get_window().get_native_window());
-        auto state = glfwGetMouseButton(window, static_cast<int32_t>(button));
+        auto state = glfwGetMouseButton(window, static_cast<int>(button));
         return state == GLFW_PRESS;
     }
 
-    std::pair<float, float> Input::get_mouse_position() {
+    std::pair<double, double> Input::get_mouse_position() {
         auto window = static_cast<GLFWwindow *>(Instance::get_instance().get_window().get_native_window());
         double x_Position, y_Position;
         glfwGetCursorPos(window, & x_Position, & y_Position);
-        return { (float)x_Position, (float)y_Position };
+        return { x_Position, y_Position };
     }
 
-    float Input::get_mouse_x() {
+    double Input::get_mouse_x() {
         auto[x, y] = get_mouse_position();
         return x;
     }
 
-    float Input::get_mouse_y() {
+    double Input::get_mouse_y() {
         auto[x, y] = get_mouse_position();
         return y;
     }
